The get_token function moved to utils file. all checked with the login and working
