+++ server/api/utils.py
@@ -8,8 +8,9 @@ from flask import abort, make_response, jsonify
 # https://stackoverflow.com/questions/21638922/custom-error-message-json-object-with-flask-restful
 from werkzeug.exceptions import HTTPException
 from app.extensions import logger
-
-
+from flask import request, current_app, abort, make_response, jsonify
+import jwt
+from datetime import datetime, timedelta
 
 error_mapper = {
     500: 'Unexpected server exception',
@@ -28,7 +29,6 @@ def exceptions_mapper(status, suffix=''):
     return status, error_mapper[status] + suffix
 
 
-
 def plt_to_src(fig):
     io = BytesIO()
     fig.savefig(io, format='png', bbox_inches='tight', pad_inches=0)
@@ -36,6 +36,7 @@ def plt_to_src(fig):
     data = base64.encodebytes(io.getvalue()).decode('utf-8')
     html_img_src = f'data:image/png;base64,{data}'
     return html_img_src
+
 
 def json_abort(status_code, message="", exception=None):
     if exception:
@@ -62,3 +63,30 @@ def json_abort(status_code, message="", exception=None):
     # We always want here call to abort
     abort(make_response(jsonify({"message": message, "code": status_code}), status_code))
     # abort(make_response(jsonify(message=message), status_code))
+
+
+def token_required(f):
+    from models.users import User
+    def wrapper(*args, **kwargs):
+        try:
+            auth_header = request.headers.get('Authorization')
+            access_token = auth_header.split(' ')[1]
+            token = jwt.decode(access_token, current_app.config['SECRET_KEY'], algorithms=["HS256"])
+            curr_user = User.query.get(token['uid'])
+            if not curr_user:
+                raise jwt.InvalidTokenError("Member not found")
+
+        except Exception as e:
+            json_abort(401, "Invalid/Expired token", e)
+        return f(*args, **kwargs, curr_user=curr_user)
+
+    wrapper.__doc__ = f.__doc__
+    wrapper.__name__ = f.__name__
+    return wrapper
+
+def get_tokens(user_id, app):
+    _access_token = jwt.encode({'uid': user_id,
+                                'exp': datetime.utcnow() + timedelta(minutes=720),
+                                'iat': datetime.utcnow()},
+                               app.config['SECRET_KEY'], algorithm='HS256').decode('utf-8')
+    return _access_token
